+++ app/src/main/java/com/mooner/starlight/ui/plugins/PluginsListAdapter.kt
@@ -15,9 +15,9 @@ import com.mooner.starlight.R
 import com.mooner.starlight.plugincore.plugin.Plugin
 import com.mooner.starlight.plugincore.plugin.StarlightPlugin
 import com.mooner.starlight.ui.plugins.config.PluginConfigActivity
+import com.mooner.starlight.ui.plugins.info.startPluginInfoActivity
 import com.mooner.starlight.ui.presets.ExpandableCardView
 import com.mooner.starlight.utils.Utils.Companion.trimLength
-import com.mooner.starlight.utils.startPluginInfoActivity
 import java.io.File
 
 class PluginsListAdapter(
