--- app/src/main/res/layout/activity_plugin_info.xml
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.plugins.info.PluginInfoActivity">
-
-<androidx.core.widget.NestedScrollView
-        android:id="@+id/scroll"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/background">
-
-    <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-        <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:gravity="center_vertical"
-                android:orientation="vertical">
-
-            <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="30dp"
-                    android:layout_marginTop="80dp"
-                    android:fontFamily="@font/nanumsquare_round_bold"
-                    android:text="정보"
-                    android:textColor="#4C4C4C"
-                    android:textSize="28sp"/>
-
-            <TextView
-                    android:id="@+id/pluginName"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="34dp"
-                    android:layout_marginBottom="30dp"
-                    android:elevation="1dp"
-                    android:fontFamily="@font/nanumsquare_round_regular"
-                    android:text="PLUGIN_NAME"
-                    android:textColor="#4C4C4C"
-                    android:textSize="14sp"
-                    app:layout_collapseMode="parallax"/>
-
-        </LinearLayout>
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
-
-            <androidx.recyclerview.widget.RecyclerView
-                    android:id="@+id/recyclerView"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:layout_marginTop="25dp"
-                    android:overScrollMode="never"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" android:layout_marginStart="8dp"
-                    android:layout_marginEnd="8dp" android:layout_marginBottom="8dp"/>
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </LinearLayout>
-</androidx.core.widget.NestedScrollView>
-
-<ImageView
-        android:id="@+id/imageViewLogo"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:layout_gravity="top|end"
-        android:scaleType="fitStart"
-        app:layout_collapseMode="parallax"
-        app:srcCompat="@drawable/svg_top_title_sunset"/>
-
-<ImageView
-        android:id="@+id/leave"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_marginStart="10dp"
-        android:layout_marginTop="10dp"
-        android:layout_marginBottom="10dp"
-        android:src="@drawable/ic_round_arrow_left_24"/>
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
